Moves the contents of admin and firebaseConfig into index, its a monster but we can now successfully createUsers while serving firebaseFunctions, renames key in config to apiKey
